+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Utils: Add NLTK's legality syllable tokenizer and sonority sequencing syllable tokenizer
 - Utils: Add NLTK's regular-expression tokenizer
 - Utils: Add Pyphen's Catalan syllable tokenizer
+- Utils: Add PyThaiNLP's ThaiSumCut
 - Utils: Add spaCy's Croatian and Swedish part-of-speech tagger
 - Utils: Add spaCy's Finnish part-of-speech tagger and lemmatizer
 - Utils: Add spaCy's Sorbian (Lower) word tokenizer and stop word list